requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
Flask-Dance==7.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
WTForms==3.0.1
python-dotenv==1.0.0
Werkzeug==2.3.7
email-validator==2.0.0
# PostgreSQL adapter - install one of these:
# psycopg2-binary==2.9.7  # For production
# psycopg2==2.9.7         # If you have PostgreSQL dev headers
# SQLite fallback (no additional requirements)
marshmallow==3.20.1
Pillow==10.3.0
requests==2.32.3
boto3==1.34.162
